wrote the construct class and associated it with the orders object.
finished up getting the membership provider to use the email as the username.

worked on the orders page for creating new constructs associated with orders.  added validators to the fields and wrote it according to david tricoli's requests on how/when data is entered. (Migrated by TFS Migration Toolkit) (Updated by TFS to TFS Connector, server: http://1x:8080/, id 1876. )

git-tfs-id: [https://tfs.caes.ucdavis.edu:8088/tfs/DefaultCollection]$/PTF/TRUNK/PTF;C1279